+++ NEWS.md
@@ -1,6 +1,7 @@
 # leprechaun 1.0.1.9000
 
 - Assets supports added for serving `type = module`
+- Add `tag2` convenience function to htmltools plugin
 
 # leprechaun 1.0.0.9000