+++ NEWS.md
@@ -7,6 +7,9 @@ the launching the app.
 - Changed `build` message when nothing is to be built
 - Added `add_app_file` function
 - Scaffold also add `_disable_autoload.R` file.
+- Keep ui file name consistent regardless of whether the app
+was created with `fluidPage` or `navbarPage`: always create
+`ui.R`.
 
 # leprechaun 1.0.0.9000